+++ package.json
@@ -45,6 +45,7 @@
     "lint:yml": "eslint \"**/*.{yml,yaml}\"",
     "lint:styles": "stylelint \"**/*.scss\"",
     "lint:lit": "yarn lit-analyzer \"src/components/**/*.tsx\"",
+    "lint:circular-imports": "madge --circular --extensions ts,tsx ./",
     "start": "storybook dev -p 6006",
     "test": "wtr --coverage --group default",
     "test:spec": "wtr --group spec",
@@ -99,6 +100,7 @@
     "jsx-dom": "^8.0.7",
     "lint-staged": "^15.0.2",
     "lit-analyzer": "^2.0.1",
+    "madge": "^6.1.0",
     "npm-run-all": "^4.1.5",
     "playwright": "^1.38.1",
     "postcss": "^8.4.28",
