+++ src/components/core/testing/index.ts
@@ -1,5 +1,4 @@
 export * from './event-spy';
 export * from './scroll';
-export * from './slotchange-events';
 export * from './wait-for-condition';
 export * from './wait-for-render';
