+++ src/components/sbb-action-group/sbb-action-group.spec.ts
@@ -1,4 +1,3 @@
-import { patchSlotchangeEvent } from '../core/testing';
 import { expect, fixture } from '@open-wc/testing';
 import { html } from 'lit/static-html.js';
 import { SbbActionGroup } from './sbb-action-group';
@@ -58,8 +57,6 @@ describe('sbb-action-group', () => {
           </sbb-link>
         </sbb-action-group>
       `)) as SbbActionGroup;
-      patchSlotchangeEvent(root);
-
       expect(assertButtons(root, (b) => b.size === 'l')).to.be.ok;
     });
 
@@ -75,8 +72,6 @@ describe('sbb-action-group', () => {
           </sbb-link>
         </sbb-action-group>
       `)) as SbbActionGroup;
-      patchSlotchangeEvent(root);
-
       expect(assertButtons(root, (b) => b.size === 'm')).to.be.ok;
     });
 
@@ -92,8 +87,6 @@ describe('sbb-action-group', () => {
           </sbb-link>
         </sbb-action-group>
       `)) as SbbActionGroup;
-      patchSlotchangeEvent(root);
-
       expect(Array.from(root.querySelectorAll('sbb-link')).every((l) => l.variant === 'block')).to
         .be.ok;
     });
@@ -110,8 +103,6 @@ describe('sbb-action-group', () => {
           </sbb-link>
         </sbb-action-group>
       `)) as SbbActionGroup;
-      patchSlotchangeEvent(root);
-
       expect(Array.from(root.querySelectorAll('sbb-link')).every((l) => l.size === 's')).to.be.ok;
     });
   });
