+++ src/components/sbb-datepicker-toggle/sbb-datepicker-toggle.tsx
@@ -6,12 +6,12 @@ import { isValidAttribute, setAttribute } from '../core/dom';
 import { documentLanguage, HandlerRepository, languageChangeHandlerAspect } from '../core/eventing';
 import { i18nShowCalendar } from '../core/i18n';
 import { SbbCalendar } from '../sbb-calendar/index';
-import type { SbbDatepicker } from '../sbb-datepicker';
 import {
   datepickerControlRegisteredEvent,
   getDatePicker,
   InputUpdateEvent,
-} from '../sbb-datepicker/sbb-datepicker.helper';
+  type SbbDatepicker,
+} from '../sbb-datepicker';
 import { SbbTooltip } from '../sbb-tooltip';
 import { SbbTooltipTrigger } from '../sbb-tooltip-trigger';
 import style from './sbb-datepicker-toggle.scss?lit&inline';
